refactor: create mergeUserDataFromRequest method to combine lender and requester in the request object
