+++ src/microservices/Request.js
@@ -3,12 +3,12 @@ import {
   basePut,
   basePost,
   baseDelete,
-  queryGet,
   basePatch,
 } from '../utils/baseRequest.js';
 import endpoint from '../utils/endpoint.js';
+import { mergeUserDataFromRequest } from '../utils/mergeUser.js';
 
-const { request, user } = endpoint[process.env.NODE_ENV] || endpoint.fallback;
+const { request } = endpoint[process.env.NODE_ENV] || endpoint.fallback;
 
 async function getProductCategory(req, res) {
   res.json(await baseGet(`${request}/product_category`));
@@ -27,31 +27,10 @@ async function createProductCategory(req, res) {
 
 async function getRequest(req, res) {
   const response = await baseGet(`${request}/requests`);
-
   const requests = response.data.requests;
+  const mergedUserWithRequest = await mergeUserDataFromRequest(requests);
 
-  const requesters = requests.map(request => request.requester);
-  const lenders = requests.map(request => request.lender);
-
-  const usersIds = [...lenders, ...requesters];
-  const usersIdsString = usersIds.join(',');
-
-  const users = await queryGet(`${user}/users`, {
-    requestUsers: usersIdsString,
-  });
-
-  const mergedRequest = requests.map(function joinUsersDataToRequest(request) {
-    const requester = users.find(user => user.useremail === request.requester);
-    const lender = users.find(user => user.useremail === request.lender);
-
-    return {
-      ...request,
-      lender,
-      requester,
-    };
-  });
-
-  res.status(200).json(mergedRequest);
+  res.status(200).json(mergedUserWithRequest);
 }
 
 async function updateLender(req, res) {
